+++ amazon/dynamodb/client/src/main/kotlin/org/http4k/connect/amazon/dynamodb/model/Attribute.kt
@@ -141,6 +141,17 @@ class Attribute<FINAL>(
             )
         }
 
+        fun defaulted(name: String, default: Attribute<OUT>, description: String? = null): Attribute<OUT> {
+            val getLens = get(name)
+            val setLens = set(name)
+            return Attribute(
+                dataType,
+                Meta(false, location, paramMeta, name, description),
+                { getLens(it).run { if (isEmpty()) default(it) else first() } },
+                { out: OUT?, target -> setLens(out?.let { listOf(it) } ?: emptyList(), target) }
+            )
+        }
+
         override fun defaulted(name: String, default: OUT, description: String?): Attribute<OUT> {
             val getLens = get(name)
             val setLens = set(name)