+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.15.3.0
+- **http4k-connect-*** - Upgrade dependencies.
+- **http4k-connect-amazon.dynamodb** - Support defaulted() for falling back to another column.
+
 ### v3.15.2.1
 - **http4k-connect-amazon-*** - AssumeRoleWithWebIdentity response has optional fields which are not documented. Grrrr.